pyproject.toml
[tool.poetry]
name = "dlsite-async"
version = "0.1.0"
description = "DLsite Async"
authors = ["byeonhyeok <bhrevol@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/bhrevol/dlsite-async"
repository = "https://github.com/bhrevol/dlsite-async"
documentation = "https://dlsite-async.readthedocs.io"
classifiers = [
    "Development Status :: 4 - Beta",
]

[tool.poetry.urls]
Changelog = "https://github.com/bhrevol/dlsite-async/releases"

[tool.poetry.dependencies]
python = "^3.9"
lxml = "^4.9.1"
aiohttp = "^3.8.1"
Pillow = {version = "^9.2.0", optional = true}

[tool.poetry.dev-dependencies]
Pygments = ">=2.10.0"
black = ">=21.10b0"
coverage = {extras = ["toml"], version = ">=6.2"}
darglint = ">=1.8.1"
flake8 = ">=4.0.1"
flake8-bandit = ">=2.1.2"
flake8-bugbear = ">=21.9.2"
flake8-docstrings = ">=1.6.0"
flake8-rst-docstrings = ">=0.2.5"
furo = ">=2021.11.12"
isort = ">=5.10.1"
mypy = ">=0.930"
pep8-naming = ">=0.12.1"
pre-commit = ">=2.16.0"
pre-commit-hooks = ">=4.1.0"
pytest = ">=6.2.5"
pyupgrade = ">=2.29.1"
safety = ">=1.10.3"
sphinx = ">=4.3.2"
sphinx-autobuild = ">=2021.3.14"
typeguard = ">=2.13.3"
xdoctest = {extras = ["colors"], version = ">=0.15.10"}
myst-parser = {version = ">=0.16.1"}
lxml-stubs = "^0.4.0"
pytest-asyncio = "^0.19.0"
pytest-mock = "^3.8.2"
aioresponses = "^0.7.3"

[tool.poetry.extras]
pil = ["Pillow"]

[tool.poetry.group.dev.dependencies]
types-Pillow = "^9.2.2"

[tool.bandit]
skips = ["B320"]

[tool.coverage.paths]
source = ["src", "*/site-packages"]
tests = ["tests", "*/tests"]

[tool.coverage.run]
branch = true
source = ["dlsite_async", "tests"]

[tool.coverage.report]
show_missing = true
fail_under = 80

[tool.isort]
profile = "black"
multi_line_output = 3
lines_after_imports = 2

[tool.mypy]
strict = true
warn_unreachable = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
